Use int in VM, since it's enough

+++ vm/profile.c
@@ -1,11 +1,10 @@
 #ifdef ICVM_PROFILE
 
 #include <stdio.h>
-#include <inttypes.h>
 
 #include "profile.h"
 
-static int64_t max_mem_size = 0;
+static int max_mem_size = 0;
 static int inst_count = 0;
 
 void profile_init() {
@@ -13,7 +12,7 @@ void profile_init() {
     inst_count = 0;
 }
 
-void profile_addr(int64_t addr) {
+void profile_addr(int addr) {
     if (addr >= max_mem_size) {
         max_mem_size = addr + 1;
     }
@@ -24,11 +23,11 @@ void profile_inst() {
 }
 
 void profile_done() {
-    fprintf(stderr, "profile: max_mem_size = %" PRId64 " inst_count = %d\n",
+    fprintf(stderr, "profile: max_mem_size = %d inst_count = %d\n",
         max_mem_size, inst_count);
 
     FILE* fp = fopen("icvm_profile.csv", "at");
-    fprintf(fp, "%" PRId64 ", %d\n", max_mem_size, inst_count);
+    fprintf(fp, "%d, %d\n", max_mem_size, inst_count);
     fclose(fp);
 }
